Remove unused code.
