+++ code/src/components/TodoList.js
@@ -4,6 +4,9 @@ import { useSelector } from 'react-redux';
 import { TodoItem } from './TodoItem';
 import { SectionTodoList } from '../styledComponents/styled_components';
 
+// Responsible for rendering all todo-items. TodoItems component, in turn, renders 
+// the custom checkbox, the delete-button, and isChecked state.
+// The property item is passed as prop to TodoItem component.
 export const TodoList = () => {
   // Get list info for this list form the (global) redux store
   const list = useSelector((store) => store.todos.list) 
@@ -14,10 +17,10 @@ export const TodoList = () => {
     <SectionTodoList className='todo-list'>
       {list.items.map((item) => (
         <TodoItem 
-          item-text={item.text}
+          /*item-text={item.text}*/
           item={item} 
           key={item.id}>
-        </TodoItem>
+      </TodoItem>
       ))}
     </SectionTodoList>
   );
